requirements.txt
imageio==2.36.1
joblib==1.4.2
lazy_loader==0.4
networkx==3.4.2
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
threadpoolctl==3.5.0
tifffile==2024.9.20
